+++ CHANGELOG.md
@@ -7,3 +7,4 @@
   * See `/src/world_maps/data/culling_actor.h` for Map Actor culling
   * See `/src/world_maps/data/culling_map_model.h` for culling the map itself
   * See `/src/world_maps/data/culling_wm_cloud.h` for culling the World 7 clouds
+* Support for changing which boss is present in a world.